+++ src/main/java/com/revolut/transfer/services/TransferService.java
@@ -4,5 +4,5 @@ import com.revolut.transfer.exceptions.AccountNotFoundException;
 import com.revolut.transfer.exceptions.BalanceNotEnoughException;
 
 public interface TransferService {
-    void transfer(final Long fromAccountNumber, final Long toAccountNumber, final Double amount) throws AccountNotFoundException, BalanceNotEnoughException;
+    void transfer(final Long fromAccountNumber, final Long toAccountNumber, final Long amount) throws AccountNotFoundException, BalanceNotEnoughException;
 }
